requirements.txt
pan-os-python==1.12.2
pan-python==0.17.0
setuptools==80.4.0
tabulate==0.9.0
